Extract ExcelDownloadButton for easier maintenance
