+++ front/app/components/FormSync/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  downloadExcelTemplate: {
+    id: 'app.containers.AdminPage.ProjectIdeaForm.downloadExcelTemplate',
+    defaultMessage: 'Download an Excel template',
+  },
+});
