+++ front/app/containers/Admin/projects/project/inputImporter/ReviewSection/EmptyState.tsx
@@ -6,23 +6,15 @@ import {
   Text,
   colors,
   stylingConsts,
-  Button,
 } from '@citizenlab/cl2-component-library';
-import { saveAs } from 'file-saver';
 import { useParams } from 'react-router-dom';
 
 import usePhase from 'api/phases/usePhase';
 
-import useFeatureFlag from 'hooks/useFeatureFlag';
-
-import buttonMessages from 'containers/Admin/projects/project/inputForm/messages';
-import { API_PATH } from 'containers/App/constants';
-
 import DownloadPDFButtonWithModal from 'components/FormBuilder/components/FormBuilderTopBar/DownloadPDFButtonWithModal';
-import UpsellTooltip from 'components/UpsellTooltip';
+import ExcelDownloadButton from 'components/FormSync/ExcelDownloadButton';
 
 import { FormattedMessage } from 'utils/cl-intl';
-import { requestBlob } from 'utils/requestBlob';
 
 import sharedMessages from '../messages';
 
@@ -37,20 +29,7 @@ const EmptyState = () => {
 
   const participationMethod = phase?.data.attributes.participation_method;
 
-  const inputImporterAllowed = useFeatureFlag({
-    name: 'input_importer',
-    onlyCheckAllowed: true,
-  });
-
   const pdfImportSupported = isPDFUploadSupported(participationMethod);
-
-  const downloadExampleXlsxFile = async () => {
-    const blob = await requestBlob(
-      `${API_PATH}/phases/${phaseId}/importer/export_form/idea/xlsx`,
-      'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-    );
-    saveAs(blob, 'example.xlsx');
-  };
 
   return (
     <Box
@@ -95,16 +74,7 @@ const EmptyState = () => {
               phaseId={phaseId}
             />
           )}
-          <UpsellTooltip disabled={inputImporterAllowed}>
-            <Button
-              buttonStyle="secondary-outlined"
-              icon="download"
-              onClick={downloadExampleXlsxFile}
-              disabled={!inputImporterAllowed}
-            >
-              <FormattedMessage {...buttonMessages.downloadExcelTemplate} />
-            </Button>
-          </UpsellTooltip>
+          <ExcelDownloadButton phaseId={phaseId} />
         </Box>
       </Box>
     </Box>
